+++ backend/app/routers/__init__.py
@@ -1,0 +1,4 @@
+from . import users, admin
+
+__all__ = ["users", "admin"]
+
